+++ packages/plugin-esbuild/src/index.ts
@@ -6,7 +6,12 @@ export default (api: IApi) => {
     key: 'esbuild',
     config: {
       schema(joi) {
-        return joi.object();
+        return joi.object({
+          target: joi.alternatives(
+            joi.string(),
+            joi.array().items(joi.string()),
+          ),
+        });
       },
     },
     enableBy: api.EnableBy.config,
@@ -14,14 +19,17 @@ export default (api: IApi) => {
 
   api.modifyBundleConfig((memo, { type }) => {
     if (memo.optimization) {
+      const { target = 'es2015', pure } = api.config.esbuild || {};
       const optsMap = {
         [BundlerConfigType.csr]: {
-          target: 'es2015',
           minify: true,
+          target,
+          pure,
         },
         [BundlerConfigType.ssr]: {
           target: 'node10',
           minify: true,
+          pure,
         },
       };
       const opts = optsMap[type] || optsMap[BundlerConfigType.csr];