+++ packages/plugin-openapi/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/umijs/plugins/compare/@umijs/plugin-openapi@1.2.0...@umijs/plugin-openapi@1.2.1) (2021-03-26)
+
+### Bug Fixes
+
+- **onpenapi:** openapi support node10 ([#567](https://github.com/umijs/plugins/issues/567)) ([99c5eb0](https://github.com/umijs/plugins/commit/99c5eb06ab92c32a658e6470667f2914a5500965))
+
 # [1.2.0](https://github.com/umijs/plugins/compare/@umijs/plugin-openapi@1.1.0...@umijs/plugin-openapi@1.2.0) (2021-03-15)
 
 ### Features